requirements.txt
gunicorn==19.7.1
django-heroku==0.3.1
Django==3.2
django-allauth==0.44.0
django-autoslug-iplweb==1.9.5
django-cors-headers==3.7.0
django-dash==0.5.6
django-filter==2.3.0
django-nine==0.2.3
django-rest-auth==0.9.5
django-templated-mail==1.1.1
django-tinymce==3.1.0
django-username-email==2.4.2
django-versatileimagefield==2.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
PyJWT==1.7.1
Faker==8.10.0
martor==1.6.3